fix delete file undefiled, css input file

+++ backend/routes/postDel.js
@@ -1,12 +1,19 @@
 import express from 'express';
 import ArticleModel from '../models/articles';
 import fs from 'fs';
+import find from 'find';
 
 const router = express.Router();
 
 router.post('/', (req, res) => {
     ArticleModel.findById(req.body.ID, (err, art) => {
-        fs.unlink('public/images/' + art.image, (err) => {})
+        find.file('public/images/', (files) => {
+            for(let i = 0; i < files.length; i++){
+                if(files[i] === 'public/images/' + art.image){
+                    fs.unlink('public/images/' + art.image, (err) => {})
+                }
+            }
+        })
     })
     ArticleModel.findByIdAndRemove(req.body.ID, (err) => {
         if (err) {
